gcp/vertex_ai/btc_forecasting/trainer/requirements.txt
matplotlib
pandas
scikit-learn
tensorflow
numpy==1.18.5
google-cloud-storage